Extract Load_Title_Screen from TD & RA
