+++ tiberiandawn/winstub.cpp
@@ -508,46 +508,3 @@ void Memory_Error_Handler(void)
     // PostQuitMessage( 0 );
     // ExitProcess(0);
 }
-
-/***********************************************************************************************
- * Load_Title_Screen -- loads the title screen into the given video buffer                     *
- *                                                                                             *
- *                                                                                             *
- *                                                                                             *
- * INPUT:    screen name                                                                       *
- *           video buffer                                                                      *
- *           ptr to buffer for palette                                                         *
- *                                                                                             *
- * OUTPUT:   Nothing                                                                           *
- *                                                                                             *
- * WARNINGS: None                                                                              *
- *                                                                                             *
- * HISTORY:                                                                                    *
- *    7/5/96 11:30AM ST : Created                                                              *
- *=============================================================================================*/
-
-#include "filepcx.h"
-void Load_Title_Screen(char* name, GraphicViewPortClass* video_page, unsigned char* palette)
-{
-    GraphicBufferClass* load_buffer;
-    const char* ext = strrchr(name, '.');
-
-    if (!strcasecmp(ext, ".PCX"))
-        load_buffer = Read_PCX_File(name, (char*)palette, NULL, 0); // Load 640x400 title
-    else if (!strcasecmp(ext, ".CPS")) {
-        /* CPS files are hardcoded to 320x200. */
-        const int width = 320;
-        const int height = 200;
-
-        load_buffer = new GraphicBufferClass(width, height, NULL, width * (height + 4));
-        Load_Uncompress(name, *load_buffer, *load_buffer, palette);
-    } else {
-        /* Invalid title screen. */
-        DBG_ERROR("Title screen file %s do not have PCX or CPS extension", name);
-    }
-
-    if (load_buffer) {
-        load_buffer->Blit(*video_page);
-        delete load_buffer;
-    }
-}
